modified data structure
main>userinterface>utils
created config.py
